removed opentrons.server package and tests.
