+++ api/src/opentrons/deck_calibration/__init__.py
@@ -61,6 +61,16 @@ x_row = 0
 y_row = 1
 z_row = 2
 
+def expected_points():
+    slot_1_lower_left,\
+        slot_3_lower_right,\
+        slot_7_upper_left = dots_set()
+
+    return {
+        '1': slot_1_lower_left,
+        '2': slot_3_lower_right,
+        '3': slot_7_upper_left}
+
 
 def position(axis, hardware, cp=None):
     """
