+++ api/tests/opentrons/conftest.py
@@ -27,10 +27,7 @@ from opentrons.legacy_api.instruments.pipette import Pipette
 from opentrons.api.routers import MainRouter
 from opentrons.api import models
 from opentrons.data_storage import database_migration
-from opentrons.server import rpc
 from opentrons import config, types
-from opentrons.server import init
-from opentrons.deck_calibration import endpoints
 from opentrons import hardware_control as hc
 from opentrons.hardware_control import adapters, API
 from opentrons.protocol_api import ProtocolContext
@@ -191,44 +188,6 @@ def old_aspiration(monkeypatch):
 
 # -----end feature flag fixtures-----------
 
-
-@pytest.mark.skipif(aionotify is None,
-                    reason="requires inotify (linux only)")
-@pytest.fixture
-async def async_server(hardware, virtual_smoothie_env, loop):
-    app = init(hardware, loop=loop)
-    yield app
-    await app.shutdown()
-
-
-@pytest.fixture
-async def async_client(async_server, loop, aiohttp_client):
-    cli = await loop.create_task(aiohttp_client(async_server))
-    endpoints.session = None
-    try:
-        yield cli
-    finally:
-        if cli.app.on_shutdown.frozen:
-            await cli.close()
-        else:
-            await async_server.shutdown()
-
-
-@pytest.fixture
-async def dc_session(request, async_server, monkeypatch, loop):
-    """
-    Mock session manager for deck calibation
-    """
-    hw = async_server['com.opentrons.hardware']
-    await hw.cache_instruments({
-        types.Mount.LEFT: None,
-        types.Mount.RIGHT: 'p300_multi_v1'})
-    ses = endpoints.SessionManager(hw)
-    endpoints.session = ses
-    monkeypatch.setattr(endpoints, 'session', ses)
-    yield ses
-
-
 @pytest.fixture(params=["dinosaur.py"])
 def protocol(request):
     try:
@@ -263,46 +222,6 @@ def session_manager(main_router):
     return main_router.session_manager
 
 
-@pytest.fixture
-def session(loop, aiohttp_client, request, main_router):
-    """
-    Create testing session. Tests using this fixture are expected
-    to have @pytest.mark.parametrize('root', [value]) decorator set.
-    If not set root will be defaulted to None
-    """
-    from aiohttp import web
-    from opentrons.server import error_middleware
-    root = None
-    try:
-        root = request.getfixturevalue('root')
-        if not root:
-            root = main_router
-        # Assume test fixture has init to attach test loop
-        root.init(loop=loop)
-    except Exception:
-        pass
-
-    app = web.Application(middlewares=[error_middleware])
-    server = rpc.RPCServer(app, root)
-    client = loop.run_until_complete(aiohttp_client(server.app))
-    socket = loop.run_until_complete(client.ws_connect('/'))
-    token = str(uuid())
-
-    async def call(**kwargs):
-        request = {
-            '$': {
-                'token': token
-            },
-        }
-        request.update(kwargs)
-        return await socket.send_json(request)
-
-    def finalizer():
-        server.shutdown()
-    request.addfinalizer(finalizer)
-    return Session(server, socket, token, call)
-
-
 def fuzzy_assert(result, expected):
     expected_re = ['.*'.join(['^'] + item + ['$']) for item in expected]
 
@@ -313,14 +232,6 @@ def fuzzy_assert(result, expected):
         assert re.compile(
             exp.lower()).match(res.lower()), "element {}: {} didn't match {}" \
             .format(idx, res, exp)
-
-
-@pytest.fixture
-def connect(session, aiohttp_client):
-    async def _connect():
-        client = await aiohttp_client(session.server.app)
-        return await client.ws_connect('/')
-    return _connect
 
 
 @pytest.fixture
